Move NodeFeeDetails into state module, make calculateAverageFeeAndDistributorShares non-exported
